Moved Parameter extension method to External.DB.Extensions
