+++ src/RhinoInside.Revit/RhinoInside.Revit.csproj
@@ -176,6 +176,7 @@
     <Compile Include="External\DB\Extensions\ElementId.cs" />
     <Compile Include="External\DB\Extensions\Face.cs" />
     <Compile Include="External\DB\Extensions\GeometryObject.cs" />
+    <Compile Include="External\DB\Extensions\Parameter.cs" />
     <Compile Include="External\DB\Extensions\XYZ.cs" />
     <Compile Include="External\DB\ParameterBinding.cs" />
     <Compile Include="External\DB\ParameterClass.cs" />
